Further refactor to make it generic
